app/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==5.0.1
annotated-types==0.5.0
APScheduler==3.8.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
comtypes==1.1.14
cryptography==36.0.2
decorator==4.4.2
docopt==0.6.2
edge-tts==6.1.1
Flask==2.0.2
frozenlist==1.3.0
gevent==21.8.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
greenlet==1.1.2
grequests==0.6.0
h11==0.13.0
httplib2==0.21.0
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.5
importlib-metadata==6.8.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.7.0.72
opplast==1.0.13
outcome==1.1.0
packaging==23.1
pandas==2.0.3
Pillow==9.0.1
pipreqs==0.4.13
playwright==1.20.0
proglog==0.1.9
protobuf==3.20.3
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pydeck==0.8.1b0
pyee==8.1.0
pygame==2.1.2
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0
referencing==0.30.0
requests==2.26.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
selenium==4.1.3
semver==3.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
streamlit==1.12.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.63.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.1
undetected-chromedriver==3.1.5.post4
uritemplate==4.1.1
urllib3==1.26.7
validators==0.20.0
vtt-to-srt3==0.2.0.0
watchdog==3.0.0
websockets==10.1
Werkzeug==2.0.2
wsproto==1.1.0
yarg==0.1.9
yarl==1.7.2
zipp==3.16.2
zope.event==4.5.0
zope.interface==5.4.0